No rights indication + bugfixes

+++ workers/snowflake.py
@@ -28,6 +28,13 @@ async def settings(message: types.Message):
     if message.chat.type == 'private':
         await message.reply(locales.string(lang, "settingspm"))
     lang = config.get_setting(message.chat.id, "Locale")
+
+    gotme = await message.bot.get_me()
+    me = await message.bot.get_chat_member(message.chat.id, gotme.id)
+    if (type(me) == types.chat_member_member.ChatMemberMember) or (types.chat_member_administrator.ChatMemberAdministrator and (me.can_restrict_members == False or me.can_delete_messages == False)):
+        await message.reply(locales.string(lang, "NoRights"))
+        return
+
     member = await message.bot.get_chat_member(message.chat.id, message.from_user.id)
     if type(member) == types.chat_member_owner.ChatMemberOwner:
         pass
@@ -44,27 +51,31 @@ async def settings(message: types.Message):
     await message.reply(locales.string(lang, "settings"), reply_markup = keyboard)
 
 @rt.message(Command("toggle"))
-async def lang(message: types.Message):
+async def toggle(message: types.Message):
     if message.chat.type == 'private':
         lang = message.from_user.language_code
         await message.reply(locales.string(lang, "settingspm"))
         return
     else:
         lang = config.get_setting(message.chat.id, "Locale")
-        member = await message.bot.get_chat_member(message.chat.id, message.from_user.id)
+        gotme = await message.bot.get_me()
+        me = await message.bot.get_chat_member(message.chat.id, gotme.id)
+        if (type(me) == types.chat_member_member.ChatMemberMember) or (types.chat_member_administrator.ChatMemberAdministrator and (me.can_restrict_members == False or me.can_delete_messages == False)):
+            await message.reply(locales.string(lang, "NoRights"))
+            return
+        
         member = await message.bot.get_chat_member(message.chat.id, message.from_user.id)
         if type(member) == types.chat_member_owner.ChatMemberOwner:
             pass
         elif type(member) == types.chat_member_administrator.ChatMemberAdministrator and member.can_restrict_members == True:
             pass
         else:
-            await message.reply(locales.string(message.from_user.language_code, "NotAdmin"))
+            await message.reply(locales.string(lang, "NotAdmin"))
             return
         newvalue = not config.get_setting(message.chat.id, 'Enabled')
         config.save_setting(message.chat.id, 'Enabled', newvalue)
         answer = locales.string(lang, f"bot{newvalue}")
         await message.reply(answer)
-        
 
 @rt.message(Command("language"))
 async def lang(message: types.Message):
@@ -93,6 +104,11 @@ async def lang(message: types.Message):
 
 @rt.chat_member(ChatMemberUpdatedFilter(IS_NOT_MEMBER >> IS_MEMBER))
 async def joined(event: types.ChatMemberUpdated):
+    gotme = await event.bot.get_me()
+    me = await event.bot.get_chat_member(event.chat.id, gotme.id)
+
+    if (type(me) == types.chat_member_member.ChatMemberMember) or (types.chat_member_administrator.ChatMemberAdministrator and (me.can_restrict_members == False or me.can_delete_messages == False)):
+        return
     if config.get_setting(event.chat.id, 'Enabled') == False:
         return
     await event.chat.ban(event.new_chat_member.user.id)
@@ -101,8 +117,15 @@ async def joined(event: types.ChatMemberUpdated):
 
 @rt.message()
 async def anything(message: types.Message):
+    lang = config.get_setting(message.chat.id, "Locale")
     if message.content_type in [ContentType.NEW_CHAT_MEMBERS, ContentType.LEFT_CHAT_MEMBER]:
         if config.get_setting(message.chat.id, "DeleteServiceMessages") == True:
+            gotme = await message.bot.get_me()
+            me = await message.bot.get_chat_member(message.chat.id, gotme.id)
+            if (type(me) == types.chat_member_member.ChatMemberMember) or (types.chat_member_administrator.ChatMemberAdministrator and (me.can_restrict_members == False or me.can_delete_messages == False)):
+                await message.reply(locales.string(lang, "NotAdmin"))
+                return
+
             await message.delete()
 
 @rt.callback_query()
@@ -119,11 +142,18 @@ async def callback(event: types.CallbackQuery):
     else:
         await event.answer(locales.string(event.from_user.language_code, "NoformatNoadmin"))
         return
+    lang = config.get_setting(event.message.chat.id, "Locale")
+
+    gotme = await event.bot.get_me()
+    me = await event.bot.get_chat_member(event.message.chat.id, gotme.id)
+
+    if (type(me) == types.chat_member_member.ChatMemberMember) or (types.chat_member_administrator.ChatMemberAdministrator and (me.can_restrict_members == False or me.can_delete_messages == False)):
+        await event.message.edit_text(locales.string(lang, "NoRights"))
+        return
     
     if event.data[:3] in cfg:
         chatid = int(event.data[3:])
         config.save_setting(chatid, cfg[event.data[:3]], not config.get_setting(chatid, cfg[event.data[:3]]))
-        lang = config.get_setting(event.message.chat.id, "Locale")
         buttons = [
             [InlineKeyboardButton(callback_data = f"ban{event.message.chat.id}", text = locales.string(lang, "BanMembers")+" "+config.checkmark(event.message.chat.id, "BanMembers"))],
             [InlineKeyboardButton(callback_data = f"ser{event.message.chat.id}", text = locales.string(lang, "DeleteServiceMessages")+" "+config.checkmark(event.message.chat.id, "DeleteServiceMessages"))],
@@ -137,7 +167,6 @@ async def callback(event: types.CallbackQuery):
             return
 
         config.save_setting(event.message.chat.id, "Locale", event.data[:2])
-        lang = config.get_setting(event.message.chat.id, "Locale")
         buttons = []
         for i in locales.existingTranslations.keys():
             if i == lang:
